Move adapter to a better folder
